Setup basic wayland functionality

+++ linuxdoom-1.10/m_misc.c
@@ -204,7 +204,7 @@ typedef struct
 {
 	char* name;
 	int* location;
-	long long int defaultvalue;
+	intptr_t defaultvalue;
 	int scantranslate; // PC scan code hack
 	int untranslated;  // lousy hack
 } default_t;
@@ -229,15 +229,15 @@ default_t defaults[] = { { "mouse_sensitivity", &mouseSensitivity, 5 },
 
 // UNIX hack, to be removed.
 #ifdef SNDSERV
-						 { "sndserver", (int*)&sndserver_filename, (long long int)"sndserver" },
+						 { "sndserver", (int*)&sndserver_filename, (intptr_t)"sndserver" },
 						 { "mb_used", &mb_used, 2 },
 #endif
 
 #endif
 
 #ifdef LINUX
-						 { "mousedev", (int*)&mousedev, (long long int)"/dev/ttyS0" },
-						 { "mousetype", (int*)&mousetype, (long long int)"microsoft" },
+						 { "mousedev", (int*)&mousedev, (intptr_t)"/dev/ttyS0" },
+						 { "mousetype", (int*)&mousetype, (intptr_t)"microsoft" },
 #endif
 
 						 { "use_mouse", &usemouse, 1 },
@@ -258,16 +258,16 @@ default_t defaults[] = { { "mouse_sensitivity", &mouseSensitivity, 5 },
 
 						 { "usegamma", &usegamma, 0 },
 
-						 { "chatmacro0", (int*)&chat_macros[0], (long long int)HUSTR_CHATMACRO0 },
-						 { "chatmacro1", (int*)&chat_macros[1], (long long int)HUSTR_CHATMACRO1 },
-						 { "chatmacro2", (int*)&chat_macros[2], (long long int)HUSTR_CHATMACRO2 },
-						 { "chatmacro3", (int*)&chat_macros[3], (long long int)HUSTR_CHATMACRO3 },
-						 { "chatmacro4", (int*)&chat_macros[4], (long long int)HUSTR_CHATMACRO4 },
-						 { "chatmacro5", (int*)&chat_macros[5], (long long int)HUSTR_CHATMACRO5 },
-						 { "chatmacro6", (int*)&chat_macros[6], (long long int)HUSTR_CHATMACRO6 },
-						 { "chatmacro7", (int*)&chat_macros[7], (long long int)HUSTR_CHATMACRO7 },
-						 { "chatmacro8", (int*)&chat_macros[8], (long long int)HUSTR_CHATMACRO8 },
-						 { "chatmacro9", (int*)&chat_macros[9], (long long int)HUSTR_CHATMACRO9 }
+						 { "chatmacro0", (int*)&chat_macros[0], (intptr_t)HUSTR_CHATMACRO0 },
+						 { "chatmacro1", (int*)&chat_macros[1], (intptr_t)HUSTR_CHATMACRO1 },
+						 { "chatmacro2", (int*)&chat_macros[2], (intptr_t)HUSTR_CHATMACRO2 },
+						 { "chatmacro3", (int*)&chat_macros[3], (intptr_t)HUSTR_CHATMACRO3 },
+						 { "chatmacro4", (int*)&chat_macros[4], (intptr_t)HUSTR_CHATMACRO4 },
+						 { "chatmacro5", (int*)&chat_macros[5], (intptr_t)HUSTR_CHATMACRO5 },
+						 { "chatmacro6", (int*)&chat_macros[6], (intptr_t)HUSTR_CHATMACRO6 },
+						 { "chatmacro7", (int*)&chat_macros[7], (intptr_t)HUSTR_CHATMACRO7 },
+						 { "chatmacro8", (int*)&chat_macros[8], (intptr_t)HUSTR_CHATMACRO8 },
+						 { "chatmacro9", (int*)&chat_macros[9], (intptr_t)HUSTR_CHATMACRO9 }
 
 };
